+++ src/js/components/input-mask.js
@@ -2,8 +2,10 @@ export function makeInputMask() {
   let inputMaskTelElements = document.querySelectorAll('.input-mask-tel');
 
   for (let inputMaskTelElement of inputMaskTelElements) {
+    const code = inputMaskTelElement.dataset.code;
+
     Inputmask({
-      'mask': '+38 (999) 999-99-99'
+      'mask': '+' + code + ' (999) 999-99-99'
     }).mask(inputMaskTelElement)
   }
 }