+++ CHANGELOG.md
@@ -1,11 +1,21 @@
 ﻿# Change Log
 
+## [0.3.0] - 2024-03-16
+ 
+- Added new mixin: grid-list.
+- Added warning about cookies.
+- Added new mouse parallax effect with data-parallax attribute.
+- Added to scss - base/classes.
+- Added date code attribute for phone in the input.
+- Classes added: is-mobile and is-desktop.
+- Bugs fixed.
+
 ## [0.2.0] - 2024-03-15
  
 - Project structure changed.
 - Styles are now divided into groups: components, layout, pages.
 - Changed sidemenu on mobile devices.
-- Bugs fixes.
+- Bugs fixed.
  
 ## [0.1.0] - 2023-09-16